backend/requirements.txt
# KRINS-Chronicle-Keeper FastAPI Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data Validation & Serialization
pydantic==2.5.0
pydantic[email]==2.5.0
email-validator==2.1.0

# HTTP Client & CORS
httpx==0.25.2
python-multipart==0.0.6

# Environment & Configuration
python-dotenv==1.0.0

# Authentication & Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# Logging & Monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Date & Time
python-dateutil==2.8.2

# Development & Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
isort==5.13.2
mypy==1.8.0

# Optional: AI/ML for enhanced intelligence features
openai==1.6.1
numpy==1.26.3
scikit-learn==1.4.0

# Optional: Vector similarity for semantic search (if using pgvector)
sentence-transformers==2.3.1